[모바일] 게시글 상세조회 페이지와 기능 추가
- 게시글 상세 조회 페이지 추가
- 게시글 상세 조회 기능 추가

+++ AndroidStudio/Univ/app/src/main/java/kr/tutorials/yguniv_anonymous_post/rest/PostsBody.kt
@@ -1,3 +1,3 @@
 package kr.tutorials.yguniv_anonymous_post.rest
 
-data class PostsBody(var id: Int, var title: String, var createDate: String, var viewCount: Int)
+data class PostsBody(var id: Long, var title: String, var createDate: String, var viewCount: Int)
